Move matching logic into own class
This prepares for adding more related logic there.
